+++ src/main/java/frc/lib/catalyst/mechanisms/ClawMechanism.java
@@ -293,6 +293,11 @@ public class ClawMechanism extends CatalystMechanism {
         return motor;
     }
 
+    @Override
+    protected CatalystMotor primaryMotorForSysId() {
+        return motor;
+    }
+
     // ===========================================
     //                  CONFIG
     // ===========================================