+++ src/main/java/frc/lib/catalyst/mechanisms/RotationalMechanism.java
@@ -491,6 +491,11 @@ public class RotationalMechanism extends CatalystMechanism {
     }
 
     /** Get the underlying motor for advanced use. */
+    @Override
+    protected CatalystMotor primaryMotorForSysId() {
+        return motor;
+    }
+
     public CatalystMotor getMotor() {
         return motor;
     }